+++ wgpu-hal/src/dx12/device.rs
@@ -1640,6 +1640,9 @@ impl crate::Device for super::Device {
             }),
             crate::ShaderInput::SpirV(_) => {
                 panic!("SPIRV_SHADER_PASSTHROUGH is not enabled for this backend")
+            },
+            crate::ShaderInput::Msl { .. } => {
+                panic!("MLS_SHADER_PASSTHROUGH is not enabled for this backend")
             }
         }
     }