+++ wgpu/src/api/shader_module.rs
@@ -236,3 +236,21 @@ pub struct ShaderModuleDescriptorSpirV<'a> {
     pub source: Cow<'a, [u32]>,
 }
 static_assertions::assert_impl_all!(ShaderModuleDescriptorSpirV<'_>: Send, Sync);
+
+/// Descriptor for a shader module given by Metal MSL source, for use with
+/// [`Device::create_shader_module_msl`].
+///
+/// This type is unique to the Rust API of `wgpu`. In the WebGPU specification,
+/// only WGSL source code strings are accepted.
+#[derive(Debug)]
+pub struct ShaderModuleDescriptorMsl<'a> {
+    /// Entrypoint.
+    pub entry_point: String,
+    /// Debug label of the shader module. This will show up in graphics debuggers for easy identification.
+    pub label: Label<'a>,
+    /// Number of workgroups in each dimension x, y and z.
+    pub num_workgroups: (u32, u32, u32),
+    /// Shader MSL source.
+    pub source: Cow<'a, str>,
+}
+static_assertions::assert_impl_all!(ShaderModuleDescriptorMsl<'_>: Send, Sync);